Move vqsr code out
