+++ joint_calling/utils.py
@@ -25,6 +25,26 @@ logger.setLevel('INFO')
 DEFAULT_REF = 'GRCh38'
 
 
+DATAPROC_PACKAGES = [
+    'joint-calling',
+    'click',
+    'cpg-gnomad',
+    'google',
+    'slackclient',
+    'fsspec',
+    'sklearn',
+    'gcloud',
+]
+
+
+GATK_VERSION = '4.2.0.0'
+GATK_DOCKER = f'us.gcr.io/broad-gatk/gatk:{GATK_VERSION}'
+# GnarlyGenotyper is in Beta and crashes with NullPointerException when using the
+# official GATK docker, that's why using a separate image for it:
+GNARLY_DOCKER = 'gcr.io/broad-dsde-methods/gnarly_genotyper:hail_ukbb_300K'
+DRIVER_IMAGE = 'australia-southeast1-docker.pkg.dev/analysis-runner/images/driver'
+
+
 def init_hail(name: str, local_tmp_dir: str = None):
     """
     Initialize Hail and set up the directory for logs
